Moving EmailWithAttachments.ts
