+++ lib/lambda/_lib/EmailWithAttachments.ts
